Introspect like a supervisor
* supervisor:which_children/1
* supervisor:count_children/1
* sys:get_status/1,2 with supervisor callback entry

Accepting processes not included in which_children/1 as sys calls
would timeout if trying to upgrade in appup. This is equivalent to
blocking in init/1 or starting between the suspend and code change.

Accepting processes included in workers/supervisors in
count_children/1 to show accepting processes but only appear in
active once accepted successfully.