Add grace acceptor spec for graceful closing of connections
Acceptor spec #{grace => timeout()} will wait for the grace timeout in
milliseconds before before the shutdown for the spec. This allows
connections to close gracefully before being shutdown when the
application is stopped or there is a failure elsewhere.

Listen sockets should be positioned in the supervision tree so that
their owners are shutdown before the acceptor_pool. Connections can
then monitor the listen socket and perform a graceful shutdown when
the listen socket exits.